Moves non-API logging classes in internal packages
