+++ org.scala-ide.sdt.core/src/org/scalaide/core/internal/logging/log4j/Log4JConfig.scala
@@ -1,4 +1,4 @@
-package org.scalaide.logging.log4j
+package org.scalaide.core.internal.logging.log4j
 
 import org.apache.log4j.{ Level => Log4JLevel }
 import org.apache.log4j.RollingFileAppender
